requirements-minimal.txt
# Core Web Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug>=3.0.0
jinja2==3.1.2

# Database
mysql-connector-python==8.2.0
SQLAlchemy>=2.0.23
Flask-SQLAlchemy>=3.1.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Security
cryptography==41.0.7
bcrypt==4.1.2
PyJWT==2.8.0
python-jose>=3.3.0
passlib>=1.7.4

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0
email-validator==2.0.0

# Logging & Monitoring
colorlog==6.8.0
psutil==5.9.6

# Cache & Queue
redis>=4.6.0
celery>=5.3.0

# Testing (optional for production)
pytest==7.4.3
pytest-cov==4.1.0